+++ src/config/tests.rs
@@ -6,15 +6,11 @@ use std::str::FromStr;
 use std::sync::{Arc, Condvar, Mutex};
 
 use fvm_shared::address::Address;
-use fvm_shared::econ::TokenAmount;
 use indoc::formatdoc;
 use ipc_sdk::subnet_id::{SubnetID, ROOTNET_ID};
 use tempfile::NamedTempFile;
 use url::Url;
 
-use crate::config::deserialize::{
-    deserialize_subnet_id_from_map, deserialize_token_amount_from_str,
-};
 use crate::config::{Config, ReloadableConfig};
 
 // Arguments for the config's fields
@@ -189,75 +185,4 @@ fn read_config() -> Config {
     );
 
     Config::from_toml_str(config_str.as_str()).unwrap()
-}
-
-#[test]
-fn test_subnet_from_map() {
-    use serde::Deserialize;
-
-    #[derive(Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    struct SubnetIdWrapper {
-        #[allow(dead_code)]
-        #[serde(deserialize_with = "deserialize_subnet_id_from_map")]
-        id: SubnetID,
-    }
-
-    let raw_str = r#"
-    {
-        "Id": {
-            "Parent": "/root/f01",
-            "Actor": "f064"
-        }
-    }"#;
-
-    let w: Result<SubnetIdWrapper, _> = serde_json::from_str(raw_str);
-    assert!(w.is_ok());
-    assert_eq!(w.unwrap().id, SubnetID::from_str("/root/f01/f064").unwrap())
-}
-
-#[test]
-fn test_subnet_from_map_error() {
-    use serde::Deserialize;
-
-    #[derive(Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    struct SubnetIdWrapper {
-        #[allow(dead_code)]
-        #[serde(deserialize_with = "deserialize_subnet_id_from_map")]
-        id: SubnetID,
-    }
-
-    let raw_str = r#"
-    {
-        "Id": {
-            "Parent": 65,
-            "Actor": "f064"
-        }
-    }"#;
-
-    let w: Result<SubnetIdWrapper, _> = serde_json::from_str(raw_str);
-    assert!(w.is_err());
-}
-
-#[test]
-fn test_token_amount_from_str() {
-    use serde::Deserialize;
-
-    #[derive(Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    struct Wrapper {
-        #[allow(dead_code)]
-        #[serde(deserialize_with = "deserialize_token_amount_from_str")]
-        token_amount: TokenAmount,
-    }
-
-    let raw_str = r#"
-    {
-        "TokenAmount": "1000000000000000000"
-    }"#;
-
-    let w: Result<Wrapper, _> = serde_json::from_str(raw_str);
-    assert!(w.is_ok());
-    assert_eq!(w.unwrap().token_amount, TokenAmount::from_whole(1));
 }
