+++ src/lotus/message/tests.rs
@@ -1,0 +1,80 @@
+// Copyright 2022-2023 Protocol Labs
+// SPDX-License-Identifier: MIT
+use std::str::FromStr;
+
+use crate::lotus::message::deserialize::{
+    deserialize_subnet_id_from_map, deserialize_token_amount_from_str,
+};
+use fvm_shared::econ::TokenAmount;
+use ipc_sdk::subnet_id::SubnetID;
+
+#[test]
+fn test_subnet_from_map() {
+    use serde::Deserialize;
+
+    #[derive(Deserialize)]
+    #[serde(rename_all = "PascalCase")]
+    struct SubnetIdWrapper {
+        #[allow(dead_code)]
+        #[serde(deserialize_with = "deserialize_subnet_id_from_map")]
+        id: SubnetID,
+    }
+
+    let raw_str = r#"
+    {
+        "Id": {
+            "Parent": "/root/f01",
+            "Actor": "f064"
+        }
+    }"#;
+
+    let w: Result<SubnetIdWrapper, _> = serde_json::from_str(raw_str);
+    assert!(w.is_ok());
+    assert_eq!(w.unwrap().id, SubnetID::from_str("/root/f01/f064").unwrap())
+}
+
+#[test]
+fn test_subnet_from_map_error() {
+    use serde::Deserialize;
+
+    #[derive(Deserialize)]
+    #[serde(rename_all = "PascalCase")]
+    struct SubnetIdWrapper {
+        #[allow(dead_code)]
+        #[serde(deserialize_with = "deserialize_subnet_id_from_map")]
+        id: SubnetID,
+    }
+
+    let raw_str = r#"
+    {
+        "Id": {
+            "Parent": 65,
+            "Actor": "f064"
+        }
+    }"#;
+
+    let w: Result<SubnetIdWrapper, _> = serde_json::from_str(raw_str);
+    assert!(w.is_err());
+}
+
+#[test]
+fn test_token_amount_from_str() {
+    use serde::Deserialize;
+
+    #[derive(Deserialize)]
+    #[serde(rename_all = "PascalCase")]
+    struct Wrapper {
+        #[allow(dead_code)]
+        #[serde(deserialize_with = "deserialize_token_amount_from_str")]
+        token_amount: TokenAmount,
+    }
+
+    let raw_str = r#"
+    {
+        "TokenAmount": "1000000000000000000"
+    }"#;
+
+    let w: Result<Wrapper, _> = serde_json::from_str(raw_str);
+    assert!(w.is_ok());
+    assert_eq!(w.unwrap().token_amount, TokenAmount::from_whole(1));
+}
